refactor: extract plan fetching in a dedicated class
